Formatter/ByteSizeFormatter: Allow sizes above 4GB
Use uint64_t for byte-size formatting and extend units to TB so
Windows shows correct sizes instead of clamping at 4GB.

+++ src/Dialogs/FileManager.cpp
@@ -441,7 +441,8 @@ ManagedFileListWidget::OnPaintItem(Canvas &canvas, const PixelRect rc,
                              / file.download_status.size));
     } else {
       char size[32];
-      FormatByteSize(size, ARRAY_SIZE(size), file.download_status.position);
+      FormatByteSize(size, ARRAY_SIZE(size),
+             static_cast<uint64_t>(file.download_status.position));
       text.Format("%s (%s)", _("Downloading"), size);
     }
